Use current date instead of transaction date for move
The move's date is the date on which the payment gateway transaction
was created. On an authorize-capture situation this would mean a delay
between the date the transaction was created and the date the payment
was captured.

The date and maturity date of the transaction should be the date on which
the capture happend (which is also when the move is created/posted).